services/graph-api/requirements.txt
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.29.0
